rename ray_trial
