+++ python/ray/tune/trial_runner.py
@@ -7,7 +7,7 @@ import os
 import time
 import traceback
 
-from ray.tune.ray_trial import RayTrialExecutor
+from ray.tune.ray_trial_executor import RayTrialExecutor
 from ray.tune.result import TIME_THIS_ITER_S
 from ray.tune.trial import Trial
 from ray.tune.trial_scheduler import FIFOScheduler, TrialScheduler
